LayoutTests/http/tests/security/contentSecurityPolicy/object-src-param-src-blocked2.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="Content-Security-Policy" content="object-src 'none'">
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<object type="application/x-non-existent-plugin">
    <param name="src" value="resources/alert-fail.html">
</object>
</body>
</html>
